Added filtering for courses. Made so that when the user is logged in, he stays logged in. Im using refresh token sent to the backend every time the user opens the app. With the refresh token the user gets an access token to use every functionality of the app.

+++ backend/src/main/java/com/shifterwebapp/shifter/payment/Payment.java
@@ -22,7 +22,7 @@ public class Payment {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Float amount;
+    private Double amount;
 
     private Date date;
